au_proc_getmntent: optimize reading, retry
1. Since reading /proc/self/mount is a relatively costly operation,
try to avoid it by first checking that the mount point provided is
indeed aufs mount point (check for aufs magic). This way, we will
read /proc/self/mounts only if we're dealing with aufs mount.

2. While doing massively parallel aufs mounts/umounts (256 or more
in parallel, for different mount points) I observed that sometimes
auplink fails to find a specified mountpoint in /proc/self/mounts.
Presumably this happens because the Linux kernel can't provide
continuity of the file while the mount entries are being added
and removed. So far the best solution to the problem I was able
to find is to retry reading the file -- in 100% of cases it finds
the mount from the second attempt. To be on the safe size, retry
reading twice.

[v2: drop setvbuf, move aufs_super_magic check to au_plink()]

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>